Add calcRegionStats job.
This job is added to queue in result of call from the app frontend, job
processed by the worker and result is returned to region.recalcStats
method.
