+++ controllers/systemjs.js
@@ -106,7 +106,7 @@ waitDb.then(db => {
         // calcUserStats(); Moved to session queue, see session.calcUserStats
         regionsAssignPhotos();
         regionsAssignComments();
-        calcRegionStats();
+        // calcRegionStats(); // Moved to userjobs task, see region.calcRegionStats
 
         return { message: 'All finished in ' + (Date.now() - startTime) / 1000 + 's.' };
     });
@@ -731,164 +731,6 @@ waitDb.then(db => {
         }
 
         return { message: 'Photos statistics were calculated in ' + (Date.now() - startTime) / 1000 + 's' };
-    });
-
-    saveSystemJSFunc(function calcRegionStats(cids) {
-        var startTime = Date.now();
-        var doneCounter = 0;
-        var query = {};
-        var fields = { _id: 0, cid: 1, parents: 1, photostat: 1, paintstat: 1, cstat: 1 };
-
-        if (cids && cids.length) {
-            query.cid = { $in: cids };
-        }
-
-        const queueLength = db.region_stat_queue.estimatedDocumentCount();
-
-        if (queueLength) {
-            print('Heads up, removing ' + queueLength + ' queue items');
-
-            // Delete photos stat queue first
-            db.region_stat_queue.deleteMany({});
-        }
-
-        var changeCounter = 0;
-        var changeRegionCounter = 0;
-
-        function countChangingValues(current, upcoming) {
-            var changedSomething = false;
-
-            if (!current) {
-                current = {};
-            }
-
-            for (var key in upcoming) {
-                if (upcoming[key] !== current[key]) {
-                    changeCounter++;
-                    changedSomething = true;
-                }
-            }
-
-            return changedSomething;
-        }
-
-        var count = db.regions.countDocuments(query);
-
-        print('Starting stat calculation for ' + count + ' regions');
-        db.regions.find(query, fields).sort({ cid: 1 }).forEach(region => {
-            var level = region.parents && region.parents.length || 0;
-            var regionHasChildren = db.regions.countDocuments({ parents: region.cid }) > 0;
-
-            var queryC = { del: null };
-            var queryImage = {};
-            var queryPhoto = { type: 1 };
-            var queryPaint = { type: 2 };
-            var $update = {
-                photostat: {
-                    all: 0, geo: 0, own: 0, owngeo: 0,
-                    s0: 0, s1: 0, s2: 0, s3: 0, s4: 0, s5: 0, s7: 0, s9: 0,
-                },
-                paintstat: {
-                    all: 0, geo: 0, own: 0, owngeo: 0,
-                    s0: 0, s1: 0, s2: 0, s3: 0, s4: 0, s5: 0, s7: 0, s9: 0,
-                },
-                cstat: {
-                    all: 0, del: 0,
-                    s5: 0, s7: 0, s9: 0,
-                },
-            };
-
-            queryC['r' + level] = region.cid;
-            queryImage['r' + level] = region.cid;
-            queryPhoto['r' + level] = region.cid;
-            queryPaint['r' + level] = region.cid;
-
-            // Returns array of objects with count for each image type and status value
-            // [{type: 1, count: 9, statuses: {s: 0, count: 7, s: 1, count: 2...}},...]
-            var statusesForTypes = db.photos.aggregate([
-                { $match: queryImage },
-                { $project: { _id: 0, type: 1, s: 1 } },
-                { $group: { _id: { type: '$type', status: '$s' }, scount: { $sum: 1 } } },
-                { $group: {
-                    _id: '$_id.type',
-                    statuses: { $push: { s: '$_id.status', count: '$scount' } },
-                    count: { $sum: '$scount' },
-                } },
-                { $project: { type: '$_id', statuses: 1, count: 1 } },
-                { $sort: { type: 1 } },
-            ]).toArray();
-
-            var photos;
-            var paintings;
-
-            if (statusesForTypes) {
-                photos = statusesForTypes.find(stat => stat.type === 1);
-
-                paintings = statusesForTypes.find(stat => stat.type === 2);
-            }
-
-            if (photos) {
-                $update.photostat.all = photos.count;
-                photos.statuses.forEach(status => {
-                    $update.photostat['s' + status.s] = status.count;
-                });
-                $update.photostat.geo = db.photos.countDocuments((queryPhoto.geo = { $exists: true }, queryPhoto));
-
-                if (regionHasChildren) {
-                    $update.photostat.owngeo = db.photos.countDocuments((queryPhoto['r' + (level + 1)] = null, queryPhoto));
-                    $update.photostat.own = db.photos.countDocuments((delete queryPhoto.geo, queryPhoto));
-                } else {
-                    $update.photostat.owngeo = $update.photostat.geo;
-                    $update.photostat.own = $update.photostat.all;
-                }
-            }
-
-            if (paintings) {
-                $update.paintstat.all = paintings.count;
-                paintings.statuses.forEach(status => {
-                    $update.paintstat['s' + status.s] = status.count;
-                });
-                $update.paintstat.geo = db.photos.countDocuments((queryPaint.geo = { $exists: true }, queryPaint));
-
-                if (regionHasChildren) {
-                    $update.paintstat.owngeo = db.photos.countDocuments((queryPaint['r' + (level + 1)] = null, queryPaint));
-                    $update.paintstat.own = db.photos.countDocuments((delete queryPaint.geo, queryPaint));
-                } else {
-                    $update.paintstat.owngeo = $update.paintstat.geo;
-                    $update.paintstat.own = $update.paintstat.all;
-                }
-            }
-
-            $update.cstat.s5 = db.comments.countDocuments((queryC.s = 5, queryC));
-            $update.cstat.s7 = db.comments.countDocuments((queryC.s = 7, queryC));
-            $update.cstat.s9 = db.comments.countDocuments((queryC.s = 9, queryC));
-            $update.cstat.del = db.comments.countDocuments((delete queryC.s, queryC.del = { $exists: true }, queryC));
-            $update.cstat.all = $update.cstat.s5 + $update.cstat.s7 + $update.cstat.s9 + $update.cstat.del;
-
-            db.regions.updateOne({ cid: region.cid }, { $set: $update });
-
-            var currentChangeCounter = changeCounter;
-
-            countChangingValues(region.photostat, $update.photostat);
-            countChangingValues(region.paintstat, $update.paintstat);
-            countChangingValues(region.cstat, $update.cstat);
-
-            if (changeCounter !== currentChangeCounter) {
-                changeRegionCounter++;
-            }
-
-            doneCounter++;
-
-            if (doneCounter % 100 === 0) {
-                print('Calculated stats for ' + doneCounter + ' region. Cumulative time: ' + (Date.now() - startTime) / 1000 + 's');
-            }
-        });
-
-        return {
-            valuesChanged: changeCounter,
-            regionChanged: changeRegionCounter,
-            message: 'Regions statistics were calculated for ' + doneCounter + ' regions in ' + (Date.now() - startTime) / 1000 + 's',
-        };
     });
 
     // This method removes local ips from ip_hist which might have been reported accidentally from the reverse proxy
